+++ example/src/androidMain/kotlin/com/copperleaf/forms/example/android/RenderFormPreview.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.unit.dp
 import com.copperleaf.forms.compose.form.Form
 import com.copperleaf.forms.core.vm.BasicFormViewModel
 import com.copperleaf.forms.core.vm.FormSavedStateAdapter
+import com.copperleaf.json.utils.toJsonString
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
 
 @Composable
 fun RenderFormPreview(
@@ -65,7 +65,7 @@ fun RenderFormPreview(
             )
             SelectionContainer {
                 Text(
-                    json.encodeToString(JsonElement.serializer(), vmState.originalData),
+                    vmState.originalData.toJsonString(json),
                     fontFamily = FontFamily.Monospace
                 )
             }
@@ -78,7 +78,7 @@ fun RenderFormPreview(
             )
             SelectionContainer {
                 Text(
-                    json.encodeToString(JsonElement.serializer(), vmState.updatedData),
+                    vmState.updatedData.toJsonString(json),
                     fontFamily = FontFamily.Monospace
                 )
             }