+++ example/src/jvmMain/kotlin/com/copperleaf/forms/example/desktop/PreviewFormDataStore.kt
@@ -1,6 +1,9 @@
 package com.copperleaf.forms.example.desktop
 
 import com.copperleaf.forms.core.vm.FormSavedStateAdapter
+import com.copperleaf.json.utils.parseJson
+import com.copperleaf.json.utils.toJsonString
+import kotlinx.serialization.json.JsonElement
 
 object PreviewFormDataStore {
 
@@ -8,31 +11,33 @@ object PreviewFormDataStore {
 
     fun getStoreAt(path: String): FormSavedStateAdapter.Store {
         return object : FormSavedStateAdapter.Store {
-            override suspend fun loadSchema(): String {
+            override suspend fun loadSchema(): JsonElement {
                 return PreviewFormDataStore::class.java
                     .getResourceAsStream("$path/schema.json")!!
                     .bufferedReader()
                     .readText()
+                    .parseJson()
             }
 
-            override suspend fun loadUiSchema(): String {
+            override suspend fun loadUiSchema(): JsonElement {
                 return PreviewFormDataStore::class.java
                     .getResourceAsStream("$path/uiSchema.json")!!
                     .bufferedReader()
                     .readText()
+                    .parseJson()
             }
 
-            override suspend fun loadInitialData(): String {
+            override suspend fun loadInitialData(): JsonElement {
                 return allData.getOrPut(path) {
                     PreviewFormDataStore::class.java
                         .getResourceAsStream("$path/data.json")!!
                         .bufferedReader()
                         .readText()
-                }
+                }.parseJson()
             }
 
-            override suspend fun saveUpdatedData(data: String) {
-                allData[path] = data
+            override suspend fun saveUpdatedData(data: JsonElement) {
+                allData[path] = data.toJsonString()
             }
         }
     }