+++ json-forms-core/src/commonMain/kotlin/com/copperleaf/forms/core/vm/FormSavedStateAdapter.kt
@@ -9,33 +9,31 @@ import com.copperleaf.forms.core.ui.UiSchema
 import com.copperleaf.json.schema.JsonSchema
 import kotlinx.coroutines.async
 import kotlinx.coroutines.coroutineScope
-import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonNull
 
 public class FormSavedStateAdapter(
     private val store: Store,
     private val initialState: () -> FormContract.State = { FormContract.State() },
-    private val json: Json = Json,
 ) : SavedStateAdapter<
     FormContract.Inputs,
     FormContract.Events,
     FormContract.State> {
 
     public interface Store {
-        public suspend fun loadSchema(): String
-        public suspend fun loadUiSchema(): String?
+        public suspend fun loadSchema(): JsonElement
+        public suspend fun loadUiSchema(): JsonElement?
 
-        public suspend fun loadInitialData(): String?
-        public suspend fun saveUpdatedData(data: String)
+        public suspend fun loadInitialData(): JsonElement?
+        public suspend fun saveUpdatedData(data: JsonElement)
     }
 
     override suspend fun SaveStateScope<
         FormContract.Inputs,
         FormContract.Events,
         FormContract.State>.save() {
-        saveDiff({ originalData }) { updatedData ->
-            store.saveUpdatedData(json.encodeToString(JsonElement.serializer(), updatedData))
+        saveDiff({ originalData }) {
+            store.saveUpdatedData(it)
         }
     }
 
@@ -47,17 +45,11 @@ public class FormSavedStateAdapter(
         val uiSchemaJsonStringAsync = async { store.loadUiSchema() }
         val initialDataStringAsync = async { store.loadInitialData() }
 
-        val schemaJsonString = schemaJsonStringAsync.await()
-        val uiSchemaJsonString = uiSchemaJsonStringAsync.await()
-        val initialDataString = initialDataStringAsync.await()
+        val schemaJsonElement = schemaJsonStringAsync.await()
+        val uiSchemaJsonElement = uiSchemaJsonStringAsync.await() ?: createDefaultUiSchema(schemaJsonElement)
+        val initialDataJsonElement = initialDataStringAsync.await()
 
-        val schemaJsonElement = json.decodeFromString(JsonElement.serializer(), schemaJsonString)
-        val uiSchemaJsonElement = uiSchemaJsonString
-            ?.let { json.decodeFromString(JsonElement.serializer(), it) }
-            ?: run { createDefaultUiSchema(schemaJsonElement) }
-        val initialDataJsonElement = initialDataString?.let { json.decodeFromString(JsonElement.serializer(), it) }
-
-        val schema = JsonSchema.parse(schemaJsonString)
+        val schema = JsonSchema.parse(schemaJsonElement)
         val uiSchema: UiSchema = uiSchemaJsonElement.resolveUiSchema(schemaJsonElement)
 
         initialState().copy(