+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-## 0.3.0 - 2022-07-05
+## 0.4.0 - 2022-07-16
+
+- Use already-parsed JsonElement instead of forcing the library to only work with JSON Strings
+
+## 0.3.0 - 2022-07-15
 
 - Adds code editor control
 - Connects rich text widget to form state
 - adds multi-select checkbox controls
 
-## 0.2.0 - 2022-07-05
+## 0.2.0 - 2022-07-15
 
 - Adds a few more widget types
 - Slightly improves UI